# src/main/resources/static/openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "gRPC-Based TCO Cost-Efficiency Calculator API",
    "description": "Comprehensive OpenAPI contract specifying the FinOps strategies, session operations, protocol buffer compilation, and quality attribute trade-off calculator controllers.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/cost/alb": {
      "post": {
        "tags": ["Cloud Service Cost"],
        "summary": "Calculate Application Load Balancer (ALB) monthly costs",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlbCostRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ALB cost breakdown generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceCostResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tco/effective-rps": {
      "post": {
        "tags": ["Effective Volume Metrics"],
        "summary": "Evaluate structural network and processing overhead scaling limits",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EffectiveRequestPerSecondRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Calculated effective request capacity allocations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EffectiveRequestPerSecondResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cost/unit-economics": {
      "post": {
        "tags": ["Portfolio Unit Economics"],
        "summary": "Calculate micro metrics comparing request-level expenses vs consumer ARPU constraints",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnitEconomicsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Affordability impact categorizations determined",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitEconomicsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlbCostRequest": {
        "type": "object",
        "properties": {
          "albCount": { "type": "integer", "default": 1 },
          "lcuPerHour": { "type": "number", "default": 0.0 },
          "fixedPerMonthUsd": { "type": "number", "default": 16.20 },
          "lcuPerHourUsd": { "type": "number", "default": 0.008 }
        }
      },
      "ServiceCostResponse": {
        "type": "object",
        "properties": {
          "monthlyTotalUsd": { "type": "number" },
          "breakdownLines": {
            "type": "array",
            "items": { "type": "string" }
          },
          "sessionStorageKey": { "type": "string" },
          "summaryLabel": { "type": "string" }
        }
      },
      "EffectiveRequestPerSecondRequest": {
        "type": "object",
        "properties": {
          "baselineRps": { "type": "integer" },
          "tlsEnabled": { "type": "boolean" },
          "mtlsEnabled": { "type": "boolean" },
          "tlsReconnectsPerHour": { "type": "number" },
          "oauthTokenValidationMode": { "type": "string", "enum": ["LOCAL_VALIDATION", "REMOTE_INTROSPECTION", "OPA_INTEGRATION"] }
        }
      },
      "EffectiveRequestPerSecondResponse": {
        "type": "object",
        "properties": {
          "effectiveRps": { "type": "integer" },
          "handshakeExtra": { "type": "integer" },
          "oauthExtra": { "type": "integer" },
          "rpsWasAdjusted": { "type": "boolean" },
          "breakdown": {
            "type": "array",
            "items": { "type": "string" }
          },
          "oauthPreview": { "type": "string" }
        }
      },
...
      "UnitEconomicsRequest": {
        "type": "object",
        "properties": {
          "egressTransferCostUsd": { "type": "number" },
          "cloudInfraCostUsd": { "type": "number" },
          "finopsSavingUsd": { "type": "number" },
          "effectiveRps": { "type": "number" },
          "consumerCount": { "type": "integer" },
          "revenuePerUserPerMonth": { "type": "number" }
        }
      },
      "UnitEconomicsResponse": {
        "type": "object",
        "properties": {
          "totalCostPerMonth": { "type": "number" },
          "costPerUser": { "type": "number" },
          "costPerMillionRequests": { "type": "number" },
          "affordabilityImpact": { "type": "string" }
        }
      },
...
    }
  }
}